+++ zoo-modules/radio-module/Radio.svelte
@@ -1,12 +1,10 @@
 <svelte:options tag="zoo-radio"></svelte:options>
 <div class="box">
-	<zoo-input-label class="input-label" {labeltext}>
-	</zoo-input-label>
+	<zoo-input-label {labeltext}></zoo-input-label>
 	<span class="template-slot {valid ? '' : 'error'}">
 		<slot bind:this={_templateSlot}></slot>
 	</span>
-	<zoo-input-info class="input-info" valid="{valid}" inputerrormsg="{errormsg}" infotext="{infotext}">
-	</zoo-input-info>
+	<zoo-input-info class="input-info" {valid} inputerrormsg="{errormsg}" {infotext}></zoo-input-info>
 </div>
 
 <style type='text/scss'>