+++ zoo-modules/shared-module/InputLabel.svelte
@@ -1,14 +1,12 @@
 <svelte:options tag="zoo-input-label"></svelte:options>
 {#if labeltext}
-<div class="label">
-	<span>{labeltext}</span>
-</div>
+	<div>{labeltext}</div>
 {/if}
 
 <style type='text/scss'>
 	@import "variables";
 
-	.label {
+	div {
 		font-size: $p1-size;
 		line-height: $p1-line-height;
 		font-weight: 800;